Se creo patron de factory
