+++ src/main/java/com/crui/patterns/behavioral/strategy/carrito/CarritoContexto.java
@@ -1,4 +1,4 @@
-package com.crui.patterns.behavioral.strategy;
+package com.crui.patterns.behavioral.strategy.carrito;
 
 
 import java.util.ArrayList;
