+++ src/main/java/com/crui/patterns/behavioral/strategy/carrito/DescuentoInterfaz.java
@@ -1,4 +1,4 @@
-package com.crui.patterns.behavioral.strategy;
+package com.crui.patterns.behavioral.strategy.carrito;
 
 public interface DescuentoInterfaz {
     double aplicarDescuento(double total);
