+++ src/main/java/com/crui/patterns/examples/checkout/before/EmailListener.java
@@ -1,0 +1,10 @@
+package com.crui.patterns.examples.checkout.before;
+
+
+
+public class EmailListener implements OrdenEventListener {
+    @Override
+    public void onPaid(Orden order) {
+        System.out.println("[Email] Enviando comprobante a cliente. Total: $" + order.total());
+    }
+}
