+++ src/main/java/com/crui/patterns/examples/checkout/before/MedioDePago.java
@@ -1,0 +1,5 @@
+package com.crui.patterns.examples.checkout.before;
+
+public interface MedioDePago {
+    boolean pay(double amount);
+}
