+++ src/main/java/com/crui/patterns/examples/checkout/before/MercadoPagoAPI.java
@@ -1,0 +1,10 @@
+package com.crui.patterns.examples.checkout.before;
+
+
+
+public class MercadoPagoAPI {
+    public boolean runPayment(int amountInCents) {
+        System.out.println("[MercadoPagoAPI] Procesando " + amountInCents + " centavos...");
+        return amountInCents <= 15000;
+    }
+}
