+++ src/main/java/com/crui/patterns/examples/checkout/before/OrdenEventListener.java
@@ -1,0 +1,5 @@
+package com.crui.patterns.examples.checkout.before;
+
+public interface OrdenEventListener {
+    void onPaid(Orden order);
+}
