+++ src/main/java/com/crui/patterns/examples/checkout/before/Producto.java
@@ -1,0 +1,26 @@
+package com.crui.patterns.examples.checkout.before;
+
+
+
+public class Producto {
+    private final String nombre;
+    private final double precio;
+
+    public Producto(String nombre, double precio) {
+        this.nombre = nombre;
+        this.precio = precio;
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public double getPrecio() {
+        return precio;
+    }
+
+    @Override
+    public String toString() {
+        return nombre + " ($" + precio + ")";
+    }
+}
